Split out config related code into a lib
